\encoding{latin1}
\name{drgee-package}
\alias{drgee-package}
\docType{package}
\title{
Doubly Robust Generalized Estimating Equations
}
\description{
  The main function is \code{drgee}. For convenience, the
  help functions are supplied. \code{drgeeData}
  extracts variables in a format needed in estimation. \code{obeFit},
  \code{ebeFit} and \code{drFit} performs outcome nuisance model based
  estimation, exposure nuisance model based estimation and doubly robust
  estimation respectively. \code{robVcov} calculates robust
  variance of the estimates.
}
\author{
Johan Zetterqvist, Arvid \enc{Sjölander}{Sjolander} with contributions from Alexander Ploner.
}
\keyword{models}
\keyword{regression}
\keyword{robust}
